Revert "add Pause Stop and Play"
