+++ src/Commands.ts
@@ -2,6 +2,5 @@ import { Command } from "./Command";
 import { Queue } from "./commands/Queue";
 import { Play } from "./commands/Play";
 import { Stop } from "./commands/Stop";
-import { Pause } from "./commands/Pause";
 
-export const Commands: Command[] = [Queue, Play, Stop, Pause];
+export const Commands: Command[] = [Queue, Play, Stop];
