+++ src/Common.ts
@@ -16,7 +16,6 @@ import {
   VoiceConnection,
   VoiceConnectionStatus,
   entersState,
-  AudioPlayer,
 } from "@discordjs/voice";
 import play, { SpotifyTrack } from "play-dl";
 import { URL } from "url";
@@ -126,23 +125,31 @@ interface Contract {
   volume: number;
   playing: boolean;
   nowPlaying: song | null;
-  audioPlayer: AudioPlayer | null;
-  paused: boolean;
 }
 
-async function playSong(guild: Guild, song: song): Promise<song | undefined> {
-  async function destroyConnection() {
-    serverQueue?.connection?.destroy();
-    serverQueue?.audioPlayer?.removeAllListeners();
-    logger.info("Destroying connection");
-    queue.delete(guild.id);
-  }
-
-  if (!song) {
-    destroyConnection();
+export const disconnect = async (interaction: CommandInteraction) => {
+  const serverQueue = queue.get(interaction.guildId || "");
+  if (!serverQueue) {
+    await interaction.followUp({
+      embeds: [errorEmbed("I am not connected to a voice channel!")],
+      ephemeral: true,
+    });
     return;
   }
+  serverQueue?.connection?.destroy();
+  logger.info("Destroying connection");
+  queue.delete(interaction.guildId || "");
+  interaction.followUp({
+    embeds: [
+      new MessageEmbed()
+        .setTitle("Disconnected")
+        .setDescription("Bot has cleared the queue.")
+        .setColor(defaultColor),
+    ],
+  });
+};
 
+async function playSong(guild: Guild, song: song): Promise<song | undefined> {
   const serverQueue = queue.get(guild.id);
   if (!serverQueue) {
     logger.error(
@@ -156,6 +163,17 @@ async function playSong(guild: Guild, song: song): Promise<song | undefined> {
     return;
   }
 
+  async function destroyConnection() {
+    serverQueue?.connection?.destroy();
+    logger.info("Destroying connection");
+    queue.delete(guild.id);
+  }
+
+  if (!song) {
+    destroyConnection();
+    return;
+  }
+
   let stream = await play.stream(song.url);
 
   const resource = createAudioResource(stream.stream, {
@@ -166,15 +184,13 @@ async function playSong(guild: Guild, song: song): Promise<song | undefined> {
 
   const audioPlayer = createAudioPlayer({
     behaviors: {
-      noSubscriber: NoSubscriberBehavior.Stop,
+      noSubscriber: NoSubscriberBehavior.Pause,
     },
   });
   audioPlayer.play(resource);
   serverQueue.playing = true;
   serverQueue.songs.shift();
   serverQueue.nowPlaying = song;
-  serverQueue.paused = false;
-  serverQueue.audioPlayer = audioPlayer;
 
   serverQueue.connection.subscribe(audioPlayer);
 
@@ -241,8 +257,6 @@ export const addSongToQueue = async (
       volume: 0.5,
       playing: true,
       nowPlaying: null,
-      audioPlayer: null,
-      paused: true,
     };
 
     if (!interaction.guildId) {
