+++ src/commands/Play.ts
@@ -1,4 +1,4 @@
-import { CommandInteraction, Client, MessageEmbed } from "discord.js";
+import { CommandInteraction, Client } from "discord.js";
 import { SlashCommandBuilder } from "@discordjs/builders";
 
 import logger from "../utils/logger";
@@ -8,8 +8,6 @@ import {
   getSongData,
   addSongToQueue,
   errorEmbed,
-  queue,
-  defaultColor,
 } from "../Common";
 
 export const Play: Command = {
@@ -17,53 +15,13 @@ export const Play: Command = {
     .setName("play")
     .setDescription("Plays song of the url in connected voice channel.")
     .addStringOption((option) =>
-      option.setName("url").setDescription("url of the song to be played")
+      option
+        .setName("url")
+        .setDescription("url of the song to be played")
+        .setRequired(true)
     ),
   run: async (client: Client, interaction: CommandInteraction) => {
     if (await checkVoiceChannel(interaction)) {
-      if (!interaction.options.getString("url")) {
-        if (!interaction?.guildId) {
-          await interaction.followUp({
-            embeds: [errorEmbed()],
-            ephemeral: true,
-          });
-          return;
-        }
-        const serverQueue = queue.get(interaction?.guildId);
-        if (!serverQueue) {
-          await interaction.followUp({
-            embeds: [errorEmbed("Enter the url of a song!")],
-            ephemeral: true,
-          });
-          return;
-        }
-
-        const success = serverQueue?.audioPlayer?.unpause();
-        if (!success) {
-          await interaction.followUp({
-            embeds: [errorEmbed()],
-            ephemeral: true,
-          });
-          return;
-        }
-        serverQueue.paused = !success;
-        await interaction.followUp(
-          success
-            ? {
-                embeds: [
-                  new MessageEmbed()
-                    .setTitle("Play")
-                    .setDescription("Resumed queue")
-                    .setColor(defaultColor),
-                ],
-              }
-            : {
-                embeds: [errorEmbed("Error trying to play queue.")],
-                ephemeral: true,
-              }
-        );
-        return;
-      }
       const data = await getSongData(interaction);
       if (!data) {
         await interaction.followUp({ embeds: [errorEmbed()], ephemeral: true });
